funciones GCP/make_prediction/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-storage
google-cloud-datastore
tensorflow
scikit-learn
numpy
requests
h5py